Fixed some new bug of Punching Action and implement Damage Multiplier
Fixed couple of new bugs:
- Fixed "uppercutting" when user is punching and jumping at the same
time
- Fixed "Pre-Punching" by adding a reset time to punching action

Implemented damage multiplier everytime a player gets hit in the
calculation

+++ GD4SFMLGameWorld/DataTable.hpp
@@ -29,7 +29,7 @@ struct StickmanData
 {
 	//damage multiplier is replacement for HP, 
 	//the bigger it is the easier stickman go out of bounds
-	int damageMultiplier;
+	float damageMultiplier;
 	float speed;
 	TextureID texture;
 	sf::Time punchInterval;
